+++ test/http/controllers/magic_starter_notification_controller_test.dart
@@ -203,6 +203,7 @@ void main() {
       });
 
       mockGuard = MockGuard();
+      Magic.singleton('auth', () => AuthManager());
       Auth.manager.forgetGuards();
       Auth.manager.extend('mock', (_) => mockGuard);
       Config.set('auth.defaults.guard', 'mock');