+++ test/http/controllers/magic_starter_otp_controller_test.dart
@@ -200,6 +200,7 @@ void main() {
 
       // 4. Bind MockGuard for Auth facade.
       mockGuard = MockGuard();
+      Magic.singleton('auth', () => AuthManager());
       Auth.manager.forgetGuards();
       Auth.manager.extend('mock', (_) => mockGuard);
       Config.set('auth.defaults.guard', 'mock');