+++ test/ui/layouts/magic_starter_app_layout_test.dart
@@ -208,6 +208,7 @@ void main() {
     );
 
     final mockGuard = MockGuard();
+    Magic.singleton('auth', () => AuthManager());
     Auth.manager.forgetGuards();
     Auth.manager.extend(
       'mock',