+++ test/ui/views/profile/magic_starter_profile_settings_email_verification_test.dart
@@ -230,6 +230,7 @@ void main() {
 
       // 3. Bind auth with mock guard.
       mockGuard = MockGuard();
+      Magic.singleton('auth', () => AuthManager());
       Auth.manager.forgetGuards();
       Auth.manager.extend('mock', (_) => mockGuard);
       Config.set('auth.defaults.guard', 'mock');