+++ test/ui/widgets/magic_starter_user_profile_dropdown_test.dart
@@ -80,6 +80,7 @@ void main() {
 
     // Mock Auth guard
     mockGuard = MockGuard();
+    Magic.singleton('auth', () => AuthManager());
     Auth.manager.forgetGuards();
     Auth.manager.extend('mock', (_) => mockGuard);
     Config.set('auth.defaults.guard', 'mock');